cleanup: update to cookiecutter for git releases until PyPI
