+++ scripts/requirements.txt
@@ -1,1 +1,1 @@
-cookiecutter
+git+https://github.com/cookiecutter/cookiecutter.git@master
