+++ templates/bentoml_gallery_{{ cookiecutter.__project_dir }}/bentofile.yaml
