+++ templates/bentoml_gallery_{{ cookiecutter.__project_dir }}/requirements.txt
