+++ templates/bentoml_gallery_{{ cookiecutter.__project_dir }}/tests/test_prediction.py
