+++ templates/cookiecutter.json
@@ -6,5 +6,8 @@
     "include_tests": [
         "True",
         "False"
-    ]
+    ],
+    "__project_dir": "{{ cookiecutter.framework.lower().replace('-', '_').replace(' ', '_').replace('scikit_learn','sklearn') }}",
+    "__project_slug": "{{ cookiecutter.project_name.lower().replace('-', '_').replace(' ', '_') }}",
+    "__full_name": "{{ cookiecutter.framework.lower().replace('-', '_').replace(' ', '_').replace('scikit_learn','sklearn') }}_{{ cookiecutter.project_name.lower().replace('-', '_').replace(' ', '_') }}"
 }
